requirements.txt
Django==4.0.3
psycopg2==2.9.3
djangorestframework==3.13.1
django-heroku==0.3.1
gunicorn==20.0.4
django-cors-headers==3.7.0
dj-database-url==0.5.0
python-decouple==3.4